fix(cli): keep 5-arg generate_dart_bindings stable; expose uniffi_dart::main
Restore the pre-#152 5-arg `generate_dart_bindings` as a backward-compatible
shim so downstream `uniffi-bindgen.rs` helpers (payjoin-ffi, bdk-dart) keep
compiling unchanged. The `crate_name`/`try_format` knobs move behind a
`#[non_exhaustive]` `DartBindgenOptions` (`generate_dart_bindings_with_options`),
so future options never break the signature again.

Expose the CLI as `uniffi_dart::main()` (moved from src/bin.rs into src/cli.rs,
gated on the `binary` feature) — the convention used by
`uniffi::uniffi_bindgen_main()` and `uniffi_bindgen_cs::main()` — so downstreams
can forward argv instead of calling the generator with positional args.

Gate the external `dart format` subprocess on `try_format` so `--no-format`
truly skips formatting (previously only in-tree genco formatting was gated).
