+++ src/lib.rs
@@ -7,4 +7,13 @@ pub use build::generate_scaffolding;
 
 pub mod gen;
 
+// The bindgen CLI, exposed as `uniffi_dart::main()` so downstream
+// `uniffi-bindgen.rs` helpers can forward argv to it (the same convention as
+// `uniffi::uniffi_bindgen_main()` / `uniffi_bindgen_cs::main()`), instead of
+// calling the generator with positional args.
+#[cfg(feature = "binary")]
+mod cli;
+#[cfg(feature = "binary")]
+pub use cli::main;
+
 pub use uniffi_dart_macro::*;
